fix linter

+++ pkg/storage/storage_test.go
@@ -416,7 +416,7 @@ func TestCacheResize(t *testing.T) {
 			Excl: 1,
 		},
 	}
-	vol.On("Limit", uint64(cacheSize)).Return(nil)
+	vol.On("Limit", cacheSize).Return(nil)
 	err := m.checkAndResizeCache(&vol, cacheSize)
 	require.NoError(t, err)
 
@@ -444,7 +444,7 @@ func TestCacheResize(t *testing.T) {
 			Excl: 0, // no files
 		},
 	}
-	vol.On("Limit", uint64(cacheSize)).Return(nil)
+	vol.On("Limit", cacheSize).Return(nil)
 	err = m.checkAndResizeCache(&vol, cacheSize)
 	require.NoError(t, err)
 
@@ -468,7 +468,7 @@ func TestCacheResize(t *testing.T) {
 			Excl: 91,
 		},
 	}
-	vol.On("Limit", uint64(100+cacheSize)).Return(nil)
+	vol.On("Limit", 100+cacheSize).Return(nil)
 	err = m.checkAndResizeCache(&vol, cacheSize)
 	require.NoError(t, err)
